Implement search worker for local search functionality
